More discoverable
I was trying to execute http/app by itself because I didn’t understand
how to run the example. So I added comments to app.js with a little
direction on how to get the example up and running. I was also adding
paths to the URL and the http server wouldn’t finish the req/resp
communication. So I removed the case statement in app.js so that it
responds for every request to make it more robust.